refactor(http): split protocol.rs into thin facade + per-primitive modules
Break the 633-line `crates/dcc-mcp-http/src/protocol.rs` module into
one sibling per MCP primitive so downstream maintainers can jump
straight to the file that matches the JSON-RPC method they are
inspecting:

- `protocol_jsonrpc.rs` — JSON-RPC 2.0 envelope + `error_codes`.
- `protocol_lifecycle.rs` — `initialize`, `ServerCapabilities`,
  `ClientRoot` / `RootsListResult`, `LoggingSetLevelParams`,
  `ElicitationCreate*`.
- `protocol_tools.rs` — `ListToolsResult`, `McpTool`,
  `McpToolAnnotations`, `CallTool*`, `ToolContent`.
- `protocol_resources.rs` — `McpResource`, `ReadResource*`,
  `ResourceContents`, `SubscribeResourceParams`, and the
  `RESOURCE_NOT_ENABLED_ERROR` constant.
- `protocol_prompts.rs` — `McpPrompt`, `ListPromptsResult`,
  `GetPrompt*`, `McpPromptMessage`, `McpPromptContent`.
- `protocol_sse.rs` — `format_sse_event` + cursor pagination helpers.

`protocol.rs` is now a thin facade that keeps only protocol-version
negotiation (`negotiate_protocol_version`, `MCP_PROTOCOL_VERSION`,
`SUPPORTED_PROTOCOL_VERSIONS`) and the session / header / method
constants. Every public type is re-exported via `pub use`, so no
call-site import change is required.

Documentation: `docs/guide/architecture.md` now records the new layout
under the `dcc-mcp-http` section.
