+++ crates/dcc-mcp-http/src/protocol_jsonrpc.rs
@@ -1,0 +1,132 @@
+//! JSON-RPC 2.0 envelope + standard error codes.
+//!
+//! Extracted from the original monolithic `protocol.rs` as part of
+//! the Batch B thin-facade split (`auto-improve`).
+
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct JsonRpcRequest {
+    pub jsonrpc: String,
+    pub id: Option<Value>,
+    pub method: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub params: Option<Value>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct JsonRpcResponse {
+    pub jsonrpc: String,
+    pub id: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub result: Option<Value>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<JsonRpcError>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct JsonRpcError {
+    pub code: i64,
+    pub message: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data: Option<Value>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct JsonRpcNotification {
+    pub jsonrpc: String,
+    pub method: String,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub params: Option<Value>,
+}
+
+/// A single JSON-RPC message (request, response, or notification).
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(untagged)]
+pub enum JsonRpcMessage {
+    Request(JsonRpcRequest),
+    Response(JsonRpcResponse),
+    Notification(JsonRpcNotification),
+}
+
+/// A batch of JSON-RPC messages.
+pub type JsonRpcBatch = Vec<JsonRpcMessage>;
+
+/// Standard JSON-RPC error codes.
+pub mod error_codes {
+    pub const PARSE_ERROR: i64 = -32700;
+    pub const INVALID_REQUEST: i64 = -32600;
+    pub const METHOD_NOT_FOUND: i64 = -32601;
+    pub const INVALID_PARAMS: i64 = -32602;
+    pub const INTERNAL_ERROR: i64 = -32603;
+
+    /// Issue #354 — the target tool declared capabilities that the hosting
+    /// DCC adapter did not advertise at startup. The error `data` payload
+    /// includes the `tool`, `required_capabilities`, `declared_capabilities`
+    /// and `missing_capabilities` fields so clients can react programmatically.
+    pub const CAPABILITY_MISSING: i64 = -32001;
+
+    /// Issue #354 — the client invoked a `workspace://` URI but did not
+    /// advertise any MCP `roots` on this session. The error `data` carries
+    /// the original path.
+    pub const NO_WORKSPACE_ROOTS: i64 = -32602;
+}
+
+impl JsonRpcResponse {
+    pub fn success(id: Option<Value>, result: Value) -> Self {
+        Self {
+            jsonrpc: "2.0".to_string(),
+            id,
+            result: Some(result),
+            error: None,
+        }
+    }
+
+    pub fn error(id: Option<Value>, code: i64, message: impl Into<String>) -> Self {
+        Self {
+            jsonrpc: "2.0".to_string(),
+            id,
+            result: None,
+            error: Some(JsonRpcError {
+                code,
+                message: message.into(),
+                data: None,
+            }),
+        }
+    }
+
+    /// Like [`Self::error`] but carries a structured `data` payload per
+    /// JSON-RPC 2.0 §5.1. Used by issue #354 for `capability_missing` and
+    /// `no workspace roots` so clients can machine-read the surrounding
+    /// context (missing capabilities, advertised roots, …).
+    pub fn error_with_data(
+        id: Option<Value>,
+        code: i64,
+        message: impl Into<String>,
+        data: Option<Value>,
+    ) -> Self {
+        Self {
+            jsonrpc: "2.0".to_string(),
+            id,
+            result: None,
+            error: Some(JsonRpcError {
+                code,
+                message: message.into(),
+                data,
+            }),
+        }
+    }
+
+    pub fn method_not_found(id: Option<Value>, method: &str) -> Self {
+        Self::error(
+            id,
+            error_codes::METHOD_NOT_FOUND,
+            format!("Method not found: {method}"),
+        )
+    }
+
+    pub fn internal_error(id: Option<Value>, msg: impl Into<String>) -> Self {
+        Self::error(id, error_codes::INTERNAL_ERROR, msg)
+    }
+}
